refactor: Rename InputStorage key to method to be consistent with InputExtrinsic
